+++ managers/kMeans.py
@@ -5,10 +5,10 @@ import warnings
 
 import pandas as pd
 from sklearn.feature_extraction.text import TfidfVectorizer
+from sklearn.feature_extraction import DictVectorizer
 from sklearn.cluster import KMeans
 from sklearn.metrics import silhouette_score
 from sklearn.pipeline import Pipeline, FeatureUnion
-from sklearn.preprocessing import MinMaxScaler
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.exceptions import ConvergenceWarning
 
@@ -17,7 +17,7 @@ from logger import createLog
 logger = createLog(__name__)
 
 
-class TextSelector(BaseEstimator, TransformerMixin):
+class FeatureSelector(BaseEstimator, TransformerMixin):
     def __init__(self, key):
         self.key = key
 
@@ -26,17 +26,6 @@ class TextSelector(BaseEstimator, TransformerMixin):
 
     def transform(self, X):
         return X[self.key]
-
-
-class NumberSelector(BaseEstimator, TransformerMixin):
-    def __init__(self, key):
-        self.key = key
-
-    def fit(self, X, y=None):
-        return self
-
-    def transform(self, X):
-        return X[[self.key]]
 
 
 class KMeansManager:
@@ -48,7 +37,7 @@ class KMeansManager:
     def createPipeline(self, transformers):
         pipelineComponents = {
             'place': ('place', Pipeline([
-                ('selector', TextSelector(key='place')),
+                ('selector', FeatureSelector(key='place')),
                 ('tfidf', TfidfVectorizer(
                     preprocessor=KMeansManager.pubProcessor,
                     stop_words='english',
@@ -58,7 +47,7 @@ class KMeansManager:
                 )
             ])),
             'publisher': ('publisher', Pipeline([
-                ('selector', TextSelector(key='publisher')),
+                ('selector', FeatureSelector(key='publisher')),
                 ('tfidf', TfidfVectorizer(
                     preprocessor=KMeansManager.pubProcessor,
                     stop_words='english',
@@ -68,7 +57,7 @@ class KMeansManager:
                 )
             ])),
             'edition': ('edition', Pipeline([
-                ('selector', TextSelector(key='edition')),
+                ('selector', FeatureSelector(key='edition')),
                 ('tfidf', TfidfVectorizer(
                     preprocessor=KMeansManager.pubProcessor,
                     stop_words='english',
@@ -78,16 +67,16 @@ class KMeansManager:
                 )
             ])),
             'pubDate': ('pubDate', Pipeline([
-                ('selector', NumberSelector(key='pubDate')),
-                ('scaler', MinMaxScaler())
+                ('selector', FeatureSelector(key='pubDate')),
+                ('vect', DictVectorizer())
             ]))
         }
 
         pipelineWeights = {
-            'place': 0.5,
+            'place': 1.0,
             'publisher': 1.0,
-            'edition': 0.75,
-            'pubDate': 2.0 
+            'edition': 1.0,
+            'pubDate': 1.25
         }
 
         return Pipeline([
@@ -95,7 +84,7 @@ class KMeansManager:
                 transformer_list=[pipelineComponents[t] for t in transformers],
                 transformer_weights={t: pipelineWeights[t] for t in transformers}
             )),
-            ('kmeans', KMeans(n_clusters=self.currentK, max_iter=150, n_init=5))
+            ('kmeans', KMeans(n_clusters=self.currentK, max_iter=100, n_init=3))
         ])
     
     @classmethod
@@ -118,18 +107,23 @@ class KMeansManager:
 
     def createDF(self):
         logger.info('Generating DataFrame from instance data')
-        self.df = pd.DataFrame([
-            {
-                'place': i.spatial if i.spatial else '',
-                'publisher': KMeansManager.getPublisher(i.publisher),
-                'pubDate': KMeansManager.getPubDateFloat(i.dates),
-                'edition': KMeansManager.getEditionStatement(i.has_version),
-                'uuid': i.uuid
-            }
-            for i in self.instances
-            if KMeansManager.emptyInstance(i) != False
-        ])
+
+        self.df = pd.DataFrame(columns=['place', 'publisher', 'pubDate', 'edition', 'uuid'])
+
+        for i in self.instances:
+            spatialData, dateData, publisherData = self.getInstanceData(i)
+
+            if bool(spatialData) or dateData != {} or publisherData:
+                self.df = self.df.append({
+                    'place': spatialData or '',
+                    'publisher': publisherData,
+                    'pubDate': dateData,
+                    'edition': self.getEditionStatement(i.has_version),
+                    'uuid': i.uuid
+                }, ignore_index=True)
+
         self.maxK = len(self.df.index) if len(self.df.index) > 1 else 2
+
         if self.maxK > 5000:
             self.maxK = int(self.maxK * (1/9))
         elif self.maxK > 1000:
@@ -138,49 +132,58 @@ class KMeansManager:
             self.maxK = int(self.maxK * (3/9))
         elif self.maxK > 250:
             self.maxK = int(self.maxK * (4/9))
+
+        if self.maxK > 1000: self.maxK = 1000
     
-    @staticmethod
-    def emptyInstance(instance):
-        return bool(instance.spatial or\
-            KMeansManager.getPubDateFloat(instance.dates) or\
-            KMeansManager.getPublisher(instance.publisher))
+    @classmethod
+    def getInstanceData(cls, instance):
+        spatial = instance.spatial
+        date = cls.getPubDateObject(instance.dates)
+        publisher = cls.getPublisher(instance.publisher)
+
+        return (spatial, date, publisher)
 
     @classmethod
-    def getPubDateFloat(cls, dates):
+    def getPubDateObject(cls, dates):
+        if dates is None or len(dates) < 1: return {}
+
         pubYears = {}
-        if not dates:
-            return 0
+
         for d in dates:
             try:
                 date, dateType = tuple(d.split('|'))
-            except ValueError:
-                return 0
-            dateStr = str(date).strip('];:.,- ')
-            if re.match(r'[0-9]{4}-[0-9]{4}', dateStr):
-                rangeMatches = re.match(r'([0-9]{4})-([0-9]{4})', dateStr)
-                startYear = rangeMatches.group(1)
-                endYear = rangeMatches.group(2)
-                pubYears[dateType] = (int(startYear) + int(endYear)) / 2
-            elif re.match(r'[0-9]{4}-[0-9]{2}-[0-9]{2}', dateStr):
-                try:
+
+                dateGroups = re.search(r'([\d\-\?]+)', date)
+
+                dateStr = dateGroups.group(1)
+
+                startYear, endYear = ('', '')
+
+                if re.match(r'[0-9]{4}-[0-9]{4}', dateStr):
+                    rangeMatches = re.match(r'([0-9]{4})-([0-9]{4})', dateStr)
+                    startYear = rangeMatches.group(1)
+                    endYear = rangeMatches.group(2)
+                elif re.match(r'[0-9]{4}-[0-9]{2}-[0-9]{2}', dateStr):
                     year, _, _ = tuple(dateStr.split('-'))
-                    pubYears[dateType] = int(year)
-                except ValueError:
-                    logger.warning('Unable to parse date {}'.format(dateStr))
-            else:
-                try:
-                    dateInt = int(dateStr)
-                    pubYears[dateType] = dateInt
-                except ValueError:
-                    logger.warning('Unable to parse date {}'.format(dateStr))
-                    pass
+                    startYear, endYear = (year, year)
+                elif re.match(r'[0-9]{4}-[0-9]{2}', dateStr):
+                    year, _ = tuple(dateStr.split('-'))
+                    startYear, endYear = (year, year)
+                else:
+                    startYear, endYear = (dateStr, dateStr)
+
+                yearParser = YearObject(startYear, endYear)
+                yearParser.setYearComponents()
+                pubYears[dateType] = dict(yearParser)
+            except (ValueError, AttributeError, IndexError) as e:
+                logger.warning('Unable to parse date {}'.format(d))
 
         for datePref in ['copyright_date', 'publication_date']:
             if datePref in pubYears.keys():
                 return pubYears[datePref]
         
         logger.debug('Unable to locate publication date')
-        return 0
+        return {}
     
     @classmethod
     def getPublisher(cls, pubStr):
@@ -203,7 +206,7 @@ class KMeansManager:
         return ''
     
     def generateClusters(self):
-        print('Generating Clusters from instances')
+        logger.info('Generating Clusters from instances')
         try:
             logger.info('Calculating number of clusters, max {}'.format(
                 self.maxK
@@ -224,7 +227,7 @@ class KMeansManager:
                 self.clusters[item].append(self.df.loc[[n]])
             except KeyError:
                 continue
-    
+
     def getK(self, start, stop):
         warnings.filterwarnings('error', category=ConvergenceWarning)
 
@@ -300,11 +303,62 @@ class KMeansManager:
             yearEds = defaultdict(list)
             logger.debug('Parsing cluster {}'.format(clust))
             for ed in self.clusters[clust]:
-                logger.debug('Adding instance to {} edition'.format(
-                    ed.iloc[0]['pubDate']
-                ))
-                yearEds[ed.iloc[0]['pubDate']].append(ed.iloc[0]['uuid'])
+                editionYear = YearObject.convertYearDictToStr(ed.iloc[0]['pubDate'])
+                logger.debug('Adding instance to {} edition'.format(editionYear))
+                yearEds[editionYear].append(ed.iloc[0]['uuid'])
             eds.extend([(year, data) for year, data in yearEds.items()])
             eds.sort(key=lambda x: x[0])
 
         return eds
+
+
+class YearObject:
+    def __init__(self, start, end):
+        self.start = str(start)
+        self.end = str(end)
+        self.century = [None, None]
+        self.decade = [None, None]
+        self.year = [None, None]
+
+    def setYearComponents(self):
+        self.setCentury()
+        self.setDecade()
+        self.setYear()
+    
+    def setCentury(self):
+        self.century[0] = int(self.start[:2])
+
+        if len(self.end) > 2: self.century[1] = int(self.end[:2])
+
+    def setDecade(self):
+        if self.start[2] not in ['-', '?']: self.decade[0] = int(self.start[2])
+
+        if len(self.end) > 2 and self.end[2] not in ['-', '?']: self.decade[1] = int(self.end[2])
+
+    def setYear(self):
+        if self.start[3] not in ['-', '?']: self.year[0] = int(self.start[3])
+
+        if len(self.end) > 2 and self.end[3] not in ['-', '?']: self.year[1] = int(self.end[3])
+
+    def __iter__(self):
+        for key in ['century', 'decade', 'year']:
+            yearComp = getattr(self, key)
+
+            if yearComp[0] is not None: yield '{}Start'.format(key), yearComp[0]
+
+            if yearComp[1] is not None: yield '{}End'.format(key), yearComp[1]
+
+    @staticmethod
+    def convertYearDictToStr(yearDict):
+        startYear = YearObject.getYearStr(yearDict, 'Start')
+        endYear = YearObject.getYearStr(yearDict, 'End')
+
+        return '{}-{}'.format(startYear, endYear) if endYear and endYear != startYear else str(startYear)
+
+    @staticmethod
+    def getYearStr(yearDict, yearType):
+        century = yearDict.get('century{}'.format(yearType), None)
+        decade = yearDict.get('decade{}'.format(yearType), None)
+        year = yearDict.get('year{}'.format(yearType), None)
+
+        return ''.join(map(lambda x: str(x) if x is not None else 'x', [century, decade, year]))