+++ managers/redis.py
@@ -16,6 +16,8 @@ class RedisManager:
 
         self.presentTime = datetime.utcnow()
         self.oneDayAgo = self.presentTime - timedelta(days=1)
+
+        self.oclcLimit = os.environ.get('OCLC_QUERY_LIMIT', 500000)
     
     def createRedisClient(self):
         self.redisClient = Redis(
@@ -41,3 +43,17 @@ class RedisManager:
             self.presentTime.strftime('%Y-%m-%dT%H:%M:%S'),
             ex=expirationTime
         )
+
+    def checkIncrementerRedis(self, service, identifier):
+        incrementValue = self.redisClient.get('{}/{}/{}'.format(
+            service, self.presentTime.strftime('%Y-%m-%d'), identifier
+        ))
+        print('REDIS TAG ', incrementValue)
+        return bool(incrementValue) and (int(incrementValue) >= self.oclcLimit)
+
+    def setIncrementerRedis(self, service, identifier):
+        redisKey = '{}/{}/{}'.format(
+            service, self.presentTime.strftime('%Y-%m-%d'), identifier
+        )
+
+        self.redisClient.incr(redisKey)